Code Refactor + Item Consume effect

+++ Assets/V1.0/Scripts/Models/Player.cs
@@ -2,8 +2,8 @@
 public class Player
 {
     public string playerName;
-    public int health;
-    public int damage;
-    public int maxHealth = 100;
+    public float health;
+    public float damage;
+    public float maxHealth = 100;
     public float moveSpeed = 3f;
 }
